Move us to Maven: Initial commit of parent pom and releng
This structure borrowed heavily from
http://git.eclipse.org/c/cdt/org.eclipse.cdt.git/tree/?h=cdt_8_8

Look at the pom.xml in the root of the repo and the releng folder.
Compare to the org.eclipse.cdt tree and notice the similarities :)

Also, notice the CROPS.setup file in releng, this is what Oomph
Installer is using to create our development environment for us.


Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>