Fix more dependencies in CROPS.setup
NOTE: if you see an error about "articial_root" failing a dependency,
this means you are missing an update-site to provide that dependency.
In other words, you have added a plugin/feature to the dependency
tree but have not added the repository from where to install it.

Or you are using a jar that is not in Orbit. Like freemarker.


Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>